[Feat] #324 전체 상품 리스트
Former-commit-id: aeab44346679ea4ce8dd632de51848eeebb31858

+++ back/src/main/java/com/msm/back/product/dto/ProductListRequestDto.java
@@ -16,5 +16,5 @@ public class ProductListRequestDto {
 
 	private Long lastId;
 
-	private Long pageSize;
+	private int pageSize;
 }
